config.yaml renamed to settings.yaml and added postfix to all files
